Extract Zif::Clickable
